refactor: rename package to tiled-catalog-broker, CLI to tcb, move notebooks to examples
- Rename src/data_catalog_service/ → src/tiled_catalog_broker/
- Package name: tiled-catalog-broker, import: tiled_catalog_broker
- CLI entry point: tcb (tcb ingest, tcb register)
- Move notebooks/ content into examples/, remove notebooks/
- Update all imports, docs, and references
