+++ examples/demo_dual_mode.py
@@ -34,7 +34,7 @@ from pathlib import Path
 # Add tiled_poc directory to path for broker package imports
 sys.path.insert(0, str(Path(__file__).parent.parent))
 
-from data_catalog_service.config import get_tiled_url, get_api_key, get_base_dir, get_service_dir
+from tiled_catalog_broker.config import get_tiled_url, get_api_key, get_base_dir, get_service_dir
 
 
 def demo_mode_a_expert(client):
@@ -46,7 +46,7 @@ def demo_mode_a_expert(client):
     - Users who want to handle file I/O themselves
     - Maximum performance (direct HDF5, no HTTP overhead)
     """
-    from data_catalog_service.query_manifest import query_manifest, load_from_manifest, build_mh_dataset
+    from tiled_catalog_broker.query_manifest import query_manifest, load_from_manifest, build_mh_dataset
 
     print("=" * 60)
     print("MODE A: Expert Path-Based Access")
