+++ examples/demo_mh_dataset.py
@@ -61,7 +61,7 @@ def _():
     sys.path.insert(0, str(Path(__file__).parent.parent))
 
     # Use config module for configuration
-    from data_catalog_service.config import get_tiled_url, get_api_key
+    from tiled_catalog_broker.config import get_tiled_url, get_api_key
 
     TILED_URL = get_tiled_url()
     API_KEY = get_api_key()
@@ -170,7 +170,7 @@ def _(mo):
 
 @app.cell
 def _(client, mo, time):
-    from data_catalog_service.query_manifest import query_manifest, load_from_manifest, build_mh_dataset
+    from tiled_catalog_broker.query_manifest import query_manifest, load_from_manifest, build_mh_dataset
 
     # Step 1: Query manifest
     _t0 = time.perf_counter()
